requirements.txt
Django==1.8.2
djangorestframework==3.2.4
Markdown==2.6.2
django-filter==0.11.0